fixes migration on grade

+++ database/migrations/2016_01_13_182649_create_grades_table.php
@@ -16,7 +16,7 @@ class CreateGradesTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned();
             $table->integer('submission_id')->unsigned();
-            $table->integer('mark')->nullable();
+            $table->float('mark')->nullable();
             $table->text('feedback')->nullable();
             $table->timestamps();
